MULTIPACKAGE this port into:     -main, the main libcanberra library     -gtk, libcanberra-gtk (gtk+2)     -gtk3 libcanberra-gtk3 (gtk+3)
This is needed for the ongoing GNOME3 work without forcing a global move
to gtk+3. The -gtk subpackage may disappear in the future when nothing
uses it.

This move will prevent the login/logout sounds to work in
gnome-session-2.X but it is a necessary evil to move forward.

EDONOTCARE/agreed by several