Rework the way we handle dependencies in the print ports: * depends on foomatic-filters when a driver installs:   share/foomatic/db/source/.../*.ppd{,.gz} * depends on foomatic-db-engine when a driver installs:   share/foomatic/db/source/.../*.xml
foomatic-db-engine now depends on foomatic-db (which itself depends on
nothing anymore).

This will ensure all the needed packages are installed for complete
foomatic-* use.
Some packages install both pre-created ppd and xml files, so theorically
it wouldn't need to depend on foomatic-db-engine, but we still do it
because we may want to regen the ppd file.

It's the best compromise I could come up with.