import seed 3.0.0
Seed is a library and interpreter, dynamically bridging (through
gobject-introspection) the WebKit JavaScriptCore engine with the GObject
type system. In a more concrete sense, Seed enables you to immediately
write applications around a significant portion of the GNOME platform,
and easily embed JavaScript as a scripting language in your GObject
library.

ok aja@