Import kguitar 0.5.1, from MAINTAINER Pascal Stumpf.
KGuitar project was started for developing an efficient and easy-to-use
environment for a guitarist. There are many tasks that guitarists do
every day and that can be made easier to do when using computer.
Examples of such tasks include finding of chord fingerings, writing &
reading tabulatures, converting classic note score to the tabulature
form and vice versa, transposition of the whole song or its parts and
many others.

with tweaks and ok jasper@