Integrate the GLib kqueue(2) backend from Dmitry Matveev's GSoC 2011.
As a consequence, explicitely disable the FAM backend.
This bring a huge improvement compared to what we've been using before
(gio-fam+gamin): file / directory monitoring and notifications are next
to instantaneous without any cpu overhead like it was before.

This means we can also drop the bootstrap FLAVOR of gamin that was
needed because of the cyclic dependency (gamin->glib2->gamin).