Update to hs-vector-0.7.1.
This is a requirement for an update of hs-snap I'm working on.

While here, add myself as maintainer (in addition to jim@, who is
too busy with private stuff for now).

There's a newer version available, but 0.7.1 is much less intrusive
to other depending ports (i.e. it doesn't break anything).

"go for it" jim@ (for co-maintaining and working on snap and related ports)