Import eliot 1.14.
Eliot is a Scrabble game, it offers 3 different game modes:
The free game mode is the standard one, ideal for having fun with your
friends or family. You can play the game against a friend so that you
can test your skills.

In the duplicate mode, all the players are faced with the same board and
letters at the same time. It is mainly used in clubs and competitions
since it takes away the chance (and tactics) component(s) of the game.

The training mode allows to play alone and make progress for the
duplicate.  It gives more freedom than the multiplayer modes.

Feedback/ok sthen@