Update to Botan 1.10.3.
A change in 1.10.2 accidentally broke ABI compatibility with
1.10.1 and earlier versions, causing programs compiled against
1.10.1 to crash if linked with 1.10.2 at runtime.

Recent versions of OpenSSL include extra information in ECC
private keys, the presence of which caused an exception when
such a key was loaded by botan. The decoding of ECC private
keys has been changed to ignore these fields if they are set.

from Brad