Import focuswriter 1.3.90.
FocusWriter is a simple, distraction-free writing environment. It
utilizes a hide-away interface that you access by moving your mouse to
the edges of the screen, allowing the program to have a familiar look
and feel to it while still getting out of the way so that you can
immerse yourself in your work.

ok sthen@