import devel/p5-Switch, ok jasper
Switch.pm provides the syntax and semantics for an explicit case
mechanism for Perl. The syntax is minimal, introducing only the
keywords "switch" and "case" and conforming to the general pattern
of existing Perl control structures. The semantics are particularly
rich, allowing any one (or more) of nearly 30 forms of matching to
be used when comparing a switch value with its various cases.

This is from CPAN and works the same as a deprecated function in
Perl base which will be removed in a future version.