Import shells/ksh93.
KSH-93 is the most recent version of the KornShell Language described
in "The KornShell Command and Programming Language," by Morris
Bolsky and David Korn of AT&T Bell Laboratories.  The KornShell is
a shell programming language, which is upward compatible with "sh"
(the Bourne Shell), and is intended to conform to the IEEE P1003.2/ISO
9945.2 Shell and Utilities standard.  KSH-93 provides an enhanced
programming environment in addition to the major command-entry
features of the BSD shell "csh".  With KSH-93, medium-sized programming
tasks can be performed at shell-level without a significant loss
in performance.  In addition, "sh" scripts can be run on KSH-93
without modification.

Distfile mirrored by aja@ (thanks!); import requested by and ok rpe@.