Import flare-0.17.1.
Flare (Free Libre Action Roleplaying Engine) is a simple game engine
built to handle a very specific kind of game: single-player 2D action
RPGs. Flare is not a reimplementation of an existing game or engine. It
is a tribute to and exploration of the action RPG genre.

Rather than building a very abstract, robust game engine, the goal of
this project is to build several real games and harvest an engine from
the common, reusable code. The first game, in progress, is a fantasy
dungeon crawl.

Flare uses simple file formats (INI style config files) for most of the
game data, allowing anyone to easily modify game contents. Open formats
are preferred (png, ogg). The game code is C++.


ok kirby@