revamp tasks a bit more: have Job::Port let each task do its own setup. as a result, try locking directly before the fork, so we can distinguish tasks that did enter the lock directly from those that have to wait.
(stuff like depend! means it has to wait/had to wait).

Also, allows tasks to just "vanish" after setup: used for checksum (the
whole code is now in there), and for junk (if several ports try to junk
simultaneously).

keep better track of the junk lock thru a job, so that it's released when
we get out of a serialized sequence of tasks (thus, no longer
release/reobtain at end of show-prepare-results/junk).

checksum prints its own distsize now, something that was awkward before.