import ports/graphics/glm, ok gonzalo@
OpenGL Mathematics (GLM) is a header only C++ mathematics library for
graphics software based on the OpenGL Shading Language (GLSL) specification.

GLM provides classes and functions designed and implemented with the same
naming conventions and functionalities as GLSL so that when a programmer
knows GLSL, he knows GLM as well, which makes it really easy to use.

This project isn't limited to GLSL features. An extension system, based
on the GLSL extension conventions, provides extended capabilities:
matrix transformations, quaternions, half-based types, random numbers,
procedural noise functions, etc...

This library works perfectly with OpenGL but it also ensures interoperability
with third party libraries and SDKs. It is a good candidate for software
rendering (Raytracing / Rasterisation), image processing, physic simulations
and any context that requires a simple and convenient mathematics library.

GLM is written in C++98 but can take advantage of C++11 when supported
by the compiler.