Import wbar-2.3.4.
Wbar is desktop application launcher that appears similiar to the dock
on an OS/X desktop.  A wbar-config gui is included although the manual
configuration is easy with entries for icon path, command and tooltip.
Wbar commonly used with lightweight, stacking window managers such as
fluxbox and openbox.

based on an original submission by J. Scott Heppler
ok jasper@