Update python 2.7 to 2.7.4.
This is a slighly different version that I sent that includes an updated
kqueue diff that works with python 2 and 3 and puts test_bsddb and
test_bsddb3 back on the skip list as they're not built outside amd64
and i386.

Only 2 tests are known to fail: test_kqueue (broken) and test_threading
(failing on 2.7.3, under investigation but likely broken). If something
else fails please send your test.log.

Tested on a bulk by landry, sparc64 by brad, ppc gonzalo and amd64 by me.