Import pushover-0.0.5, a puzzle game featuring dominoes.
Pushover is a faithful reimplementation of the game with the same name
published in 1992 by Ocean. It contains the original levels. The
graphics and sound are very similar when compared with the original
game.

The task of the game is to rearrange the dominoes on the different
platforms so that you can start a chainreaction that makes all dominoes
topple over.

various tweaks/corrections from naddy@ landry@ bcallah@

ok naddy@