Update e2fsprogs to 1.42.7, and rename the library back to libuuid now that ossp-uuid (devel/uuid) has been changed to use the alternative name libossp-uuid as used on various other OS. When third party software wants a uuid library, is is nearly always the one from e2fsprogs, so let's avoid patching for the common case.
e2fsprogs has its own libcom_err (slightly different API to heimdal);
previously we installed this under an alternative name; in this update
I have changed to statically linking this instead as it is not used
externally. The et/com_err.h header is no longer installed to avoid
ports picking it up in preference to the proper heimdal header for
this library in kerberosV/com_err.h.

Fallout will be handled in the next couple of commits.

ok ajacoutot@