Import connectagram 1.1.1, looks good to sthen@
Connectagram is a word unscrambling game. The board consists of several
scrambled words that are joined together. You can choose the length of
the words, the amount of words, and the pattern that the words are
arranged in. The game provides a hint option for times when you are
stuck, and features an online word lookup that fetches the definitions
of each word from Wiktionary. Your current progress is automatically
saved.