the tidy php extension tries to call tidyReleaseDate() to get some silly version string, but the library symbol doesnt exist for it to resolve. this points it at tidyVersion instead.
you can reproduce the problem this solves by enabling the tidy
extension and running "php-5.3 -i" or by writing a script that calls
phpinfo(). they should segfault after ld complains about a missing
symbol.

the only caveat to this is the meaning of tidyReleaseDate doesnt
map well to what tidyVersion returns now (which is a version number),
but i dont care.

ok jasper@