Mark as BROKEN until I get the time to build new bootstrappers after the 64 bit time_t switch.
If anyone wants to beat me, grab ~kili/ghc-bootstrap.sh from cvs
(or http://openbsd.dead-parrot.de/ghc-bootstrap.sh), *read and try
to understand it*, downgrade the lang/ghc port to the last 7.4.
version (for example, cvs -q up -dP -D 2013-07-09) and then run
ghc-boostrap.sh (but be aware that you'll need a lot of time, disk
space, memory, and alcohol).

This may work or not (it worked for 5.3-current back in may). If
it doesn't work, either wait or go using some real programming
language.