Import celt07-0.7.1.
The CELT codec is an experimental audio codec for use in low-delay
speech and audio communication. It's meant to close the gap between
Vorbis and Speex for applications where both high quality audio and
low delay are desired.

This package contains the celt 0.7 API.

A requirement for the upcoming audio/mumble port.