Import wizznic-0.9.9.
Wizznic is a brick-matching puzzle-game, an improved version of Puzznic.
The challenge is to clear each level of bricks by moving the bricks next
to each other, this sounds a lot easier than it is. The bricks are
heavy, so you can only push them, not lift them up.

ok landry@