Update to rrdtool 1.4.8. Merge p5-RRD into the main package.
Note that this now uses cairo for graphics generation and pango/fontconfig
for text rendering; if you are using this in a chroot jail (for cgi/php
scripts, etc) you will need to take additional steps to install the relevant
files. A script is provided to copy the relevant libraries and support files;
see /usr/local/share/doc/pkg-readmes/rrdtool-1.4.8 for more details.