Import games/flightgear.  ok landry@, "reads ok" sthen@
The goal of the FlightGear project is to create a sophisticated and open
flight simulator framework for use in research or academic environments,
pilot training, as an industry engineering tool, for DIY-ers to pursue
their favorite interesting flight simulation idea, and last but
certainly not least as a fun, realistic, and challenging desktop flight
simulator. We are developing a sophisticated, open simulation framework
that can be expanded and improved upon by anyone interested in
contributing.