split2po is a command-line tool that takes the two given DocBook XML files and produces a gettext message file (PO-file) that represents the changes between them.
xml2pot is a command-line tool that produces a gettext message
template file from the DocBook XML file original-XML.

Some utilities from original package were not ported due
to lack of C runtime for Antlr 2 in our ports.

This is ex-kdesdk member. Importing as a part of KDE 4.11.4 update.
This port won't work until update is finished, thus not hooking it up.