KDE 4.11.5. Some highlights:
Three Kate plugins written in Python migrated to Python 3 and thus
are not available for the moment.

kdelibs-4.x deconflicted with kdebindings-3.x, and kde-runtime-4.x
deconflicted with kdebase-3.x. Most notable change is the renaming
of kreadconfig and kwriteconfig by appending a "4" suffix to them.
All ports in CVS and in WIP repo are checked for usage of all renamed
binaries, but anything else could be broken. Sorry.

KWin renamed in kde-workspace again, to avoid conflict with
KFourInLine.

Default dependencies that came from kde4.port.mk now have additional
restriction ",<5". KDE Frameworks won't hit CVS during current cycle
but it's better to be proactive.