import ports/textproc/uchardet, ok landry@
uchardet is a C language binding of the original C++ implementation of
the universal charset detection library by Mozilla.

uchardet is an encoding detector library, which takes a sequence of bytes
in an unknown character encoding without any additional information, and
attempts to determine the encoding of the text.

The original code of universalchardet is available at
http://lxr.mozilla.org/seamonkey/source/extensions/universalchardet/

Techniques used by universalchardet are described at
http://www.mozilla.org/projects/intl/UniversalCharsetDetection.html