Update to spidermonkey 24.2.0.
Somewhat tested on amd64/sparc64, now requires a decent compiler so will
only build anyway on i386/amd64/sparc64/powerpc with 4.8 and maaaybe
alpha/hppa via 4.6. Let's be crazy and use clang on amd64, tough shit
for mips64(el).

Required by the GNOME steamroller, with prodding from the hobbits.
Stuff depending on it (polkit/gjs) will soon be updated to work with it.