Import py-httpie 0.8.0.
HTTPie is a command line HTTP client. Its goal is to make CLI
interaction with web services as human-friendly as possible. It provides
a simple http command that allows for sending arbitrary HTTP requests
using a simple and natural syntax, and displays colorized responses.
HTTPie can be used for testing, debugging, and generally interacting
with HTTP servers.

ok benoit@