Import sysutils/rcm
The rcm suite of tools is for managing dotfiles directories. This suite is
useful for committing your rc files to a central repository to share, but it
also scales to a more complex situation such as multiple source directories
shared between computers with some host-specific or task-specific files.

Works fine on my machine, looks good to sthen@, okay benoit@.

Ported by Mike Burns, thank you for the effort!