Import uudeview, from Mikolaj Kucharski.
ok zhuk@ on a previous version.

uudeview is program at which you can throw a mess of unordered files in
any of BinHex, uuencoded, xxencoded, Base64 or yEnc formats and expect
them all to be properly decoded. These formats are frequently
encountered when sending binary files via news or mail.