Import games/liberalcrimesquad ok benoit@
Welcome to Liberal Crime Squad! The Conservatives have taken the
Executive, Legislative, and Judicial branches of government. Over time,
the Liberal laws of this nation will erode and turn the country into a
BACKWOODS YET CORPORATE NIGHTMARE. To prevent this from happening, the
Liberal Crime Squad was established. The mood of the country is
shifting, and we need to turn things around. Go out on the streets and
indoctrinate Conservative automatons. That is, let them see their True
Liberal Nature. Then arm them and send them forth to Stop Evil.
Eventually the public will fall behind us, and we will put more Elite
Liberals in government for greater justice.