Update to ezstream 0.6.0.
From mgrimm@ (upstream)
"
This update includes a SECURITY FIX for a shell command injection
vulnerability. I recommend to update.
Please take note of the NEWS file included in the tarball (or the
homepage on icecast.org after the update), as configuration files
need to be adjusted slightly.
"

So, see http://www.icecast.org/ezstream/ for details.