Import netatalk-3.1.7.
Netatalk is an OpenSource software package, that can be used to turn a
*NIX machine into an extremely high-performance and reliable file server
for Macintosh computers.
<...>

Sharing the same name as net/netatalk but with a different pkgspec so
users can choose which one to install.
Successfully tested as an Apple Time Machine backup device.

input and ok sthen@, jasper@