Update to icinga2-2.2.4. Change rc script to use "icinga2 daemon" and rc_bg to force running in the background; it was previously using "icinga2 daemon -d", but since 2.2.2 this has been hanging on OpenBSD.
Disable vfork as was done upstream on MacOS (to stop a hang there) in the
hope it may also stop a runtime hang seen occasionally here.

There are still some odd things going on and I don't think this is production-
ready on OpenBSD yet, so the warning in DESCR-main remains, but I think this
is not worse than 2.2.1.