Add a patch from OpenConnect upstream; "If we get an AC_PKT_DISCONN message with no payload, just skip printing the disconnect reason." Fixes printing junk and possibly crashing if the server goes away.
Tweak DESCR, mention the open-source server ocserv (port for this is
ready but holding in openbsd-wip until we're done with 5.7).

(Readers might be interested to note that 7.04 is probably the last
version of OpenConnect to only support Cisco's SSL-VPN protocol; the
next version will almost certainly add client support for Juniper's
protocol).