Import p5-Test-XML-0.08, a new test dependency of p5-TAP-Formatter-JUnit.
This module provides tools for testing code which deals with XML.
These tools are compatible with the Test::More and Test::Simple modules.
At present, there is generic XML comparison support, as well as modules
for dealing with XML::SAX, XML::Twig and XML::XPath handlers.

Missing commit first spotted by sthen@ with sqlports.
This is expected to fix the portroach breakage reported by ajacoutot@;
sorry for that!

OK sthen@