Import py-libnacl-1.4.3.
This library is used to gain direct access to the functions exposed by Daniel J.
Bernstein’s nacl library via libsodium or tweetnacl. It has been constructed to
maintain extensive documentation on how to use nacl as well as being completely
portable. The file in libnacl/__init__.py can be pulled out and placed directly
in any project to give a single file binding to all of nacl.

ok jasper@