Update to tdb-1.2.12
Bite the bullet and switch to an upstream version that uses waf*.  The
plan is to use databases/tdb and devel/libtalloc in the samba4 port, and
thus to keep those ports in sync with whatever bundled version comes
with the samba4 tarball.

* trying to make sense of the waf patches is highly discouraged

While here:
- reorder the Makefile
- update the license marker
- take maintainership (discussed with nicm@, former MAINTAINER)

This is a SHARED_ONLY port, and a python wrapper is now provided.