Import libdiscid-0.6.1, a MusicBrainz DiscIDs library.  ok sthen@
Port submitted by Donovan Watteau (MAINTAINER), needed for an upcoming
audio/cmus update.

pkg/DESCR:
libdiscid is a C library for creating MusicBrainz DiscIDs from audio
CDs.  It reads a CD's table of contents (TOC) and generates an
identifier which can be used to lookup the CD at MusicBrainz
(http://musicbrainz.org).  Additionally, it provides a submission URL
for adding the DiscID to the database.

The library also provides FreeDB disc IDs, and MCN + ISRCs, if
available.

The interface of this library is new, but the DiscID algorithm and the
operating system dependent CD-ROM/DVD-ROM access code have been ported
from libmusicbrainz version 2.