Update to a newer version of the radeondrm firmware from the linux-firmware git repository.
The firmware with newer style headers is not installed as it is named
chipname_rlc.bin and we install the old style firmware named
CHIPNAME_rlc.bin as radeon-chipname_rlc.

This is not a problem now or even with a 3.14 radeondrm but will
be a flag day at some point.

ok sthen@