SECURITY update to PuTTY 0.66
"This release fixes a security hole in the terminal emulation code.
Writing a particular escape sequence to the screen in a PuTTY terminal
session could cause the terminal code to read *and potentially write*
memory outside its own data structures. This might be exploitable, so
everybody should upgrade to a fixed version."