import ports/devel/py-pbr, updated from a port by viq, ok jasper
PBR is a library that injects some useful and sensible default
behaviors into your setuptools run. It started off life as the chunks
of code that were copied between all of the OpenStack projects. Around
the time that OpenStack hit 18 different projects each with at least
3 active branches, it seems like a good time to make that code into
a proper re-usable library.