Import ports/sysutils/fwa. ok jca@
fwa - simple file change notifier keeping unix philosophy in mind.  It is
intended to be a clean and simple application to watch file changes on systems
using kqueue.  It does not execute any commands, just prints out the changed
file's name.  To handle the events, read fwa's output and act as you wish.