Noto is a font family designed to cover the entire range of the Unicode standard.
When text is rendered by a computer, sometimes characters are displayed as
"tofu". They are little boxes to indicate your device doesn't have a font to
display the text. Google has been developing a font family called Noto,
which aims to support all languages with a harmonious look and feel. Noto is
Google's answer to tofu. The name noto is to convey the idea that Google's
goal is to see "no more tofu". Noto has multiple styles and weights, and
freely available to all.

collaboration + ok czarkoff@